created sticky component
