+++ next-client/src/components/wrappers/index.ts
@@ -1,0 +1,1 @@
+export * from "./sticky/Sticky";
